libhdhomerun: bump 20220303 -> 20231109
fetch from github instead of their mirror (rationale: it lags behind,
eg: 20231109 is not available there yet)

add a patch to modify the darwin portion of their Makefile to not
attempt to build an aarch64/x86_64 cross-arch library -- we don't care
about this on NixOS and their explicit `--target' calls breaks with
our implicit `-mmacos-' in CFLAGS, where our value wins. So without
this patch, both libraries end up targeting the same architecture,
resulting in a lipo error that they are for the same arch and can't be
combined.